requirements.txt
deeppavlov
gensim
fasttext
lru-dict
methodtools
nltk
numpy
tensorflow
